+++ src/game/server/tf/bot/tf_bot_locomotion.h
@@ -37,6 +37,8 @@ public:
 	//
 	virtual void Jump( void ) OVERRIDE;								// initiate a simple undirected jump in the air
 
+	virtual bool CanClimbFakeLadders( void ) const OVERRIDE { return true; }
+
 protected:
 	virtual void AdjustPosture( const Vector &moveGoal ) { }	// never crouch to navigate
 };