+++ CHANGELOG.md
@@ -28,7 +28,7 @@ nav_generate_auto, nav_generate_auto_view_distance, tf_mvm_popfile_requested,
 mp_restartblock, tf_roundstarttalk_disable, tf_gamemode_override, 
 cl_default_networking_off, nav_save_compressed, tf_bot_spells, 
 tf_vision_force, tf_bot_buy_upgrades, tf_mirrormode, tf_revives_enable, 
-sv_mapentities_override, sv_mapentities_mod  
+sv_mapentities_override, sv_mapentities_mod, nav_begin_ladder  
 - New inputs for tf_gamerules:  
 -- SoloSaveData, SoloUnlockItem<string>, SoloUnlockItemID<int>, SoloAddCredits<int>
 - The item schema is now loaded from items_custom.txt, allowing for extension and modularity
@@ -51,6 +51,7 @@ sv_mapentities_override, sv_mapentities_mod
 - Bots can now use spellbook spells and canteens
 - Bots can now stun Merasmus and seek him out while he's hiding
 - Bots can now use all ZI abilities
+- Bots can now wall climb in VSH using nav mesh ladders
 - Bots can now use Sweeping Charge in VSH
 - Bots can now buyback in MvM
 - Medic bots can now revive teammates if revives are active in MvM/PvP